[feat] #49 proto datastore에 token 및 isLogin 필드 구현
