+++ app/src/main/proto/user_prefs.proto
@@ -5,7 +5,9 @@ option java_multiple_files = true;
 
 message UserPreferences {
     string accessToken = 1;
-    OnboardingInfo onboardingInfo = 2;
+    string refreshToken = 2;
+    bool isLogin = 3;
+    OnboardingInfo onboardingInfo = 4;
 }
 
 message OnboardingInfo {
